requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
-e git+https://github.com/henrymoss/BOSS.git@f19eaf7231ed007cce9e12fba0f7f936eb48cfdb#egg=boss
cairocffi==1.3.0
cffi==1.15.0
chime==0.6.0
click==8.0.3
cvxpy==1.1.15
cycler==0.10.0
Cython==3.0.0a9
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
ecos==2.0.7.post1
emcee==3.1.1
emukit==0.4.9
entrypoints==0.3
GPy==1.10.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.5
notebook==6.4.6
numpy==1.21.2
osqp==0.6.2.post0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
paramz==0.9.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycparser==2.20
pyDOE==0.3.8
Pygments==2.10.0
pymdptoolbox==4.0b3
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-igraph==0.9.6
pytz==2021.3
pyzmq==22.3.0
qdldl==0.1.5.post0
qtconsole==5.2.2
QtPy==1.11.3
regex==2021.11.10
SciencePlots==1.0.9
scikit-learn==1.0.2
scipy==1.7.1
scs==2.1.4
Send2Trash==1.8.0
six==1.16.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.6.0
